requirements.txt
streamlit>=1.28.0
neo4j>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
altair>=5.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0 